Finally no more separate *_mem functions! now we use a Memory type (actually a struct) that includes both address and value.

+++ instructions/move.h
@@ -1,6 +1,6 @@
 #ifndef MOVE_H
 #define MOVE_H
 #include "../includes.h"
-void movb(uint8_t pos1, uint8_t pos2);
-void movw(uint8_t pos1, uint8_t pos2);
+void movb(uint16_t pos1, uint16_t pos2);
+void movw(uint16_t pos1, uint16_t pos2);
 #endif
